Move input into a separate module
